contacts: Use doubles in blob detection
This actually speeds things up and prevents precision errors
with very small contacts.

+++ src/contacts/basic/cluster.hpp
@@ -13,21 +13,21 @@ namespace iptsd::contacts::basic {
 
 class Cluster {
 public:
-	f32 x = 0;
-	f32 y = 0;
-	f32 xx = 0;
-	f32 yy = 0;
-	f32 xy = 0;
-	f32 w = 0;
+	f64 x = 0;
+	f64 y = 0;
+	f64 xx = 0;
+	f64 yy = 0;
+	f64 xy = 0;
+	f64 w = 0;
 
 public:
 	Cluster(Heatmap &hm, index2_t center);
 
-	math::Vec2<f32> mean();
-	math::Mat2s<f32> cov();
+	math::Vec2<f64> mean();
+	math::Mat2s<f64> cov();
 
 private:
-	void add(index2_t pos, f32 val);
+	void add(index2_t pos, f64 val);
 	void check(Heatmap &hm, index2_t pos);
 };
